Day of the week - 1st project
This C program asks the user to enter a date. After that, the software will make several calculations to find which day of the week this date refers to.
The program was made in a structured and sequential way completely.
Sorry if the code is bad, it was my first job in the area, and I did my best.