Vestibular - 2nd project
This program asks the user to enter name, age, grade obtained and the course chosen by the candidate. The software checks whether the student will pass the entrance exam or not and informs it on the screen. At the end, it also informs the name and grade of the best student in each course.
This was my second project, where I practiced looping loops and string manipulation.